Move asm.xtensa into anal. fix dupplicated symbols linkage bug ##anal
