+++ libr/asm/meson.build
@@ -66,7 +66,6 @@ r_asm_sources = [
   #join_paths('p','asm_x86_vm.c'),
   join_paths('p','asm_xap.c'),
   join_paths('p','asm_xcore_cs.c'),
-  join_paths('p','asm_xtensa.c'),
   # join_paths('p','asm_z80.c'),
   join_paths('p','asm_jdh8.c'),
   #join_paths('arch','6502','6502dis.c'),
@@ -156,10 +155,6 @@ r_asm_sources = [
   join_paths('arch','vax','vax-dis.c'),
   join_paths('arch','whitespace','wsdis.c'),
   # join_paths('arch','xap','dis.c'),
-  join_paths('arch','xtensa','gnu','elf32-xtensa.c'),
-  join_paths('arch','xtensa','gnu','xtensa-dis.c'),
-  join_paths('arch','xtensa','gnu','xtensa-isa.c'),
-  join_paths('arch','xtensa','gnu','xtensa-modules.c'),
   #join_paths('arch','z80','expressions.c'),
   # join_paths('arch','z80','z80.c'),
   #join_paths('arch','z80','z80asm.c'),
